first pass through care_denoise done
